Updated MGSKosmicTask framework to 32/64 bit capable.

+++ KosmicTask/KosmicTask/source/JAProcessInfo.h
@@ -3,11 +3,11 @@
 @interface JAProcessInfo : NSObject {
 	
 @private
-    int numberOfProcesses;
+    size_t numberOfProcesses;
     NSMutableArray *processList;
 }
 - (id) init;
-- (int)numberOfProcesses;
+- (size_t)numberOfProcesses;
 - (void)obtainFreshProcessList;
 - (BOOL)findProcessWithName:(NSString *)procNameToSearch;
 - (NSArray *)descendentsOfPID:(int)parentPID;
