+++ KosmicTask/KosmicTask/source/mother framework/GSSauthenticate.h
@@ -68,7 +68,7 @@
 				GSS_S_COMPLETE is returned.  In all cases, whenever outToken has a value, it must be sent
 				to the client.
 */
-OM_uint32 AuthenticateGSS( char *inToken, int inTokenLen, char **outToken, int *outTokenLen, char **inOutServiceName, 
+OM_uint32 AuthenticateGSS( char *inToken, int inTokenLen, char **outToken, size_t *outTokenLen, char **inOutServiceName, 
 						   char **outUserPrinc, gss_ctx_id_t *inOutGSScontext, gss_cred_id_t *inOutGSScreds );
